refactor: basic types
