+++ src/crypto-providers/witnessingValidation.ts
@@ -4,7 +4,7 @@ import {
   Uint,
 } from 'cardano-hw-interop-lib'
 import { Errors } from '../errors'
-import { HwSigningData } from '../basicTypes'
+import { HwSigningData } from '../argTypes'
 import {
   SigningMode,
   SigningParameters,
